requirements.txt
xDjango==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
mysqlclient==2.2.7
PyMySQL==1.1.2